app/requirements.txt
Flask>=2.0.2
dapr
azure-identity
azure-ai-formrecognizer
Jinja2
gunicorn
jsonpath2
arrow
colorlog
jsonschema
flask-cors
SQLAlchemy
psycopg2-binary
pandas
openpyxl